refactor(#4712): de-giant turn view reconciler via pure move
Move the reconciler facade, reaction application, state resolution, and durable target store into focused child modules while preserving behavior and locking in the sub-threshold root.
