+++ ARCHITECTURE.md
@@ -946,9 +946,13 @@ src/
 │   │   │   ├── reconcile.rs
 │   │   │   └── watcher_backstop.rs
 │   │   ├── turn_view_reconciler/
+│   │   │   ├── api.rs
+│   │   │   ├── apply.rs
 │   │   │   ├── orphan_sweep.rs
 │   │   │   ├── queue_repair.rs
 │   │   │   ├── reaction_set.rs
+│   │   │   ├── resolution.rs
+│   │   │   ├── store.rs
 │   │   │   └── tests.rs
 │   │   ├── voice_barge_in/
 │   │   │   ├── tests/
